Update README.md
I am a student of English under the National University of Bangladesh, a passionate teacher with three years of teaching experience, and a dedicated humanitarian & development professional. My journey combines education, teaching, and humanitarian service with the aim of creating a positive social impact.

As a teacher, I enjoy guiding students with care and dedication. As a humanitarian, I engage in community development and social initiatives, believing in the power of education and empathy to bring change. My goal is to contribute to both academic and social development, building a career that blends knowledge, teaching, and humanitarian values.

Signed-off-by: ALI HASAN MOHAMMED MIZAN <ahmmizan23@gmail.com>